Move ArUco detector to its directory
