+++ aruco_detector/aruco_detector.py
